Refactor: reorganize
