+++ runner/bin/flutter/flutter_wrappers.dart
@@ -1,12 +1,14 @@
 
+import '../app/throw_unimplemented.dart';
+
+void registerPlugins(FlutterEngine engine) {
+  ThrowUnimplemented.throwIfOn();
+}
+
 class DartProject {
   final String sourceDir;
 
   DartProject(this.sourceDir);
-}
-
-void registerPlugins(FlutterEngine engine) {
-  throw UnimplementedError();
 }
 
 class FlutterEngine {
@@ -15,17 +17,20 @@ class FlutterEngine {
   }
 }
 class FlutterView {
-  int get nativeWindow => throw UnimplementedError();
+  int get nativeWindow {
+    ThrowUnimplemented.throwIfOn();
+    return 0;
+  }
 }
 
 class FlutterViewController {
   final int width;
   final int height;
-  late final DartProject _project;
+  late final DartProject project;
   late final FlutterEngine engine;
   late final FlutterView view;
 
-  FlutterViewController(this.width, this.height, this._project) {
+  FlutterViewController(this.width, this.height, this.project) {
     // todo: this must generate errors if not create
     engine = FlutterEngine();
     // todo: this must generate errors if not create
@@ -33,6 +38,7 @@ class FlutterViewController {
   }
 
   int wndProc(int hWnd, int uMsg, int wParam, int lParam){
-    throw UnimplementedError();
+    ThrowUnimplemented.throwIfOn();
+    return 0;
   }
 }
