+++ runner/bin/main.dart
@@ -1,10 +1,13 @@
-import 'flutter_window.dart';
-import 'flutter_wrappers.dart';
-import 'native_app.dart';
+import 'app/native_app.dart';
+import 'app/throw_unimplemented.dart';
+import 'flutter/flutter_window.dart';
+import 'flutter/flutter_wrappers.dart';
 
 void main() {
+  ThrowUnimplemented.off();
+
   final dartProject = DartProject('!!! path directory !!!');
-  final wnd = FlutterWindow(dartProject)
+  FlutterWindow(dartProject)
     ..title = 'New Window'
     ..show();
